requirements.txt
streamlit
rembg
pillow
pandas
requests
onnxruntime
pytest
